+++ R/get_groundwater_data.R
@@ -1,22 +1,3 @@
-
-#' Helper function: get groundwater options
-#'
-#' @return return available groundwater data options and prepare for being used
-#' as input for \code{\link{get_groundwater_data}}
-#' @export
-#' @examples
-#' get_groundwater_options()
-#'
-get_groundwater_options <- function ()
-{
-  overview_options <- unlist(get_overview_options())
-
-  is_groundwater <- startsWith(names(overview_options), "groundwater")
-
-  overview_options[is_groundwater] %>%
-    gsub(pattern = "gws", replacement = "gwl")
-}
-
 #' Get Groundwater Data
 #'
 #' @description wrapper function to scrape all available raw data, i.e. groundwater
@@ -43,24 +24,28 @@ get_groundwater_data <- function(
     debug = TRUE
 )
 {
+  #kwb.utils::assignPackageObjects("wasserportal")
   result <- lapply(
     X = seq_along(groundwater_options),
     FUN = function(i) {
-      gw_option <- groundwater_options[i]
+      option_key <- groundwater_options[i]
+      option_name <- names(option_key)
       kwb.utils::catAndRun(
         messageText = sprintf(
           "Importing '%s' data (%d/%d)",
-          names(gw_option), i, length(groundwater_options)
+          option_name, i, length(groundwater_options)
         ),
         expr = {
+          ids <- stations$overview_list[[option_name]]$Messstellennummer
           lapply(
-            X = stations$overview_list[[names(gw_option)]]$Messstellennummer,
-            FUN = function(station) {
+            X = ids,
+            FUN = function(id) {
               kwb.utils::catAndRun(
-                sprintf("Downloading Messstellennummer == '%s'", station),
-                expr = {
-                  read_wasserportal_raw_gw(station, stype = gw_option)
-                },
+                sprintf(
+                  "Downloading Messstellennummer '%s' (%d/%d)",
+                  id, which(id == ids), length(ids)
+                ),
+                expr = read_wasserportal_raw_gw(id, stype = option_key),
                 dbg = debug
               )
             }) %>%
@@ -70,4 +55,22 @@ get_groundwater_data <- function(
     })
 
   stats::setNames(result, names(groundwater_options))
+}
+
+#' Helper function: get groundwater options
+#'
+#' @return return available groundwater data options and prepare for being used
+#' as input for \code{\link{get_groundwater_data}}
+#' @export
+#' @examples
+#' get_groundwater_options()
+#'
+get_groundwater_options <- function ()
+{
+  overview_options <- unlist(get_overview_options())
+
+  is_groundwater <- startsWith(names(overview_options), "groundwater")
+
+  overview_options[is_groundwater] %>%
+    gsub(pattern = "gws", replacement = "gwl")
 }